chore: 플레이리스트 컴포넌트 정리
